refactor(chore): Refactor project structure splitting up domains
