+++ tap_github/commits/commits.py
@@ -1,0 +1,31 @@
+# Python imports
+# Third-Party imports
+import singer
+# Project imports
+
+
+def get_all_commits(schema, repo_path,  state, mdata, start_date):
+    '''
+    https://developer.github.com/v3/repos/commits/#list-commits-on-a-repository
+    '''
+    bookmark = get_bookmark(state, repo_path, "commits", "since", start_date)
+    if bookmark:
+        query_string = '?since={}'.format(bookmark)
+    else:
+        query_string = ''
+    with metrics.record_counter('commits') as counter:
+        for response in authed_get_all_pages(
+                'commits',
+                'https://api.github.com/repos/{}/commits{}'.format(repo_path, query_string)
+        ):
+            commits = response.json()
+            extraction_time = singer.utils.now()
+            for commit in commits:
+                commit['_sdc_repository'] = repo_path
+                with singer.Transformer() as transformer:
+                    rec = transformer.transform(commit, schema, metadata=metadata.to_map(mdata))
+                singer.write_record('commits', rec, time_extracted=extraction_time)
+                singer.write_bookmark(state, repo_path, 'commits', {'since': singer.utils.strftime(extraction_time)})
+                counter.increment()
+    return state
+
