+++ tap_github/gh_client.py
@@ -1,0 +1,227 @@
+# Python imports
+import argparse
+import os
+import json
+import collections
+import time
+import requests
+# Third-Party imports
+import singer
+import singer.bookmarks as bookmarks
+import singer.metrics as metrics
+import backoff
+# Project imports
+from streams import get_catalog
+# set default timeout of 300 seconds
+REQUEST_TIMEOUT = 300
+session = requests.Session()
+logger = singer.get_logger()
+
+
+def get_bookmark(state, repo, stream_name, bookmark_key, start_date):
+    repo_stream_dict = bookmarks.get_bookmark(state, repo, stream_name)
+    if repo_stream_dict:
+        return repo_stream_dict.get(bookmark_key)
+    if start_date:
+        return start_date
+    return None
+
+
+def translate_state(state, catalog, repositories):
+    '''
+    This tap used to only support a single repository, in which case the
+    state took the shape of:
+    {
+      "bookmarks": {
+        "commits": {
+          "since": "2018-11-14T13:21:20.700360Z"
+        }
+      }
+    }
+    The tap now supports multiple repos, so this function should be called
+    at the beginning of each run to ensure the state is translate to the
+    new format:
+    {
+      "bookmarks": {
+        "singer-io/tap-adwords": {
+          "commits": {
+            "since": "2018-11-14T13:21:20.700360Z"
+          }
+        }
+        "singer-io/tap-salesforce": {
+          "commits": {
+            "since": "2018-11-14T13:21:20.700360Z"
+          }
+        }
+      }
+    }
+    '''
+    nested_dict = lambda: collections.defaultdict(nested_dict)
+    new_state = nested_dict()
+
+    for stream in catalog['streams']:
+        stream_name = stream['tap_stream_id']
+        for repo in repositories:
+            if bookmarks.get_bookmark(state, repo, stream_name):
+                return state
+            if bookmarks.get_bookmark(state, stream_name, 'since'):
+                new_state['bookmarks'][repo][stream_name]['since'] = bookmarks.get_bookmark(state, stream_name, 'since')
+
+    return new_state
+
+
+def raise_for_error(resp, source):
+    error_code = resp.status_code
+    try:
+        response_json = resp.json()
+    except Exception:
+        response_json = {}
+
+    if error_code == 404:
+        details = ERROR_CODE_EXCEPTION_MAPPING.get(error_code).get("message")
+        if source == "teams":
+            details += ' or it is a personal account repository'
+        message = "HTTP-error-code: 404, Error: {}. Please refer \'{}\' for more details.".format(details, response_json.get("documentation_url"))
+    else:
+        message = "HTTP-error-code: {}, Error: {}".format(
+            error_code, ERROR_CODE_EXCEPTION_MAPPING.get(error_code, {}).get("message", "Unknown Error") if response_json == {} else response_json)
+
+    exc = ERROR_CODE_EXCEPTION_MAPPING.get(error_code, {}).get("raise_exception", GithubException)
+    raise exc(message) from None
+
+
+def calculate_seconds(epoch):
+    current = time.time()
+    return int(round((epoch - current), 0))
+
+
+def rate_throttling(response):
+    if int(response.headers['X-RateLimit-Remaining']) == 0:
+        seconds_to_sleep = calculate_seconds(int(response.headers['X-RateLimit-Reset']))
+
+        if seconds_to_sleep > 600:
+            message = "API rate limit exceeded, please try after {} seconds.".format(seconds_to_sleep)
+            raise RateLimitExceeded(message) from None
+
+        logger.info("API rate limit exceeded. Tap will retry the data collection after %s seconds.", seconds_to_sleep)
+        time.sleep(seconds_to_sleep)
+
+
+# pylint: disable=dangerous-default-value
+# during 'Timeout' error there is also possibility of 'ConnectionError',
+# hence added backoff for 'ConnectionError' too.
+@backoff.on_exception(backoff.expo, (requests.Timeout, requests.ConnectionError), max_tries=5, factor=2)
+def authed_get(source, url, headers={}):
+    with metrics.http_request_timer(source) as timer:
+        session.headers.update(headers)
+        resp = session.request(method='get', url=url, timeout=get_request_timeout())
+        if resp.status_code != 200:
+            raise_for_error(resp, source)
+        timer.tags[metrics.Tag.http_status_code] = resp.status_code
+        rate_throttling(resp)
+        return resp
+
+
+def authed_get_all_pages(source, url, headers={}):
+    while True:
+        r = authed_get(source, url, headers)
+        yield r
+        if 'next' in r.links:
+            url = r.links['next']['url']
+        else:
+            break
+
+
+def get_abs_path(path):
+    return os.path.join(os.path.dirname(os.path.realpath(__file__)), path)
+
+
+def generate_pr_commit_schema(commit_schema):
+    pr_commit_schema = commit_schema.copy()
+    pr_commit_schema['properties']['pr_number'] = {
+        "type":  ["null", "integer"]
+    }
+    pr_commit_schema['properties']['pr_id'] = {
+        "type": ["null", "string"]
+    }
+    pr_commit_schema['properties']['id'] = {
+        "type": ["null", "string"]
+    }
+
+    return pr_commit_schema
+
+
+def load_schemas():
+    schemas = {}
+
+    for filename in os.listdir(get_abs_path('schemas')):
+        path = get_abs_path('schemas') + '/' + filename
+        file_raw = filename.replace('.json', '')
+        with open(path) as file:
+            schemas[file_raw] = json.load(file)
+
+    schemas['pr_commits'] = generate_pr_commit_schema(schemas['commits'])
+    return schemas
+
+
+def write_metadata(mdata, values, breadcrumb):
+    mdata.append(
+        {
+            'metadata': values,
+            'breadcrumb': breadcrumb
+        }
+    )
+
+
+def populate_metadata(schema_name, schema):
+    mdata = metadata.new()
+    #mdata = metadata.write(mdata, (), 'forced-replication-method', KEY_PROPERTIES[schema_name])
+    mdata = metadata.write(mdata, (), 'table-key-properties', KEY_PROPERTIES[schema_name])
+
+    for field_name in schema['properties'].keys():
+        if field_name in KEY_PROPERTIES[schema_name]:
+            mdata = metadata.write(mdata, ('properties', field_name), 'inclusion', 'automatic')
+        else:
+            mdata = metadata.write(mdata, ('properties', field_name), 'inclusion', 'available')
+
+    return mdata
+
+
+def verify_repo_access(url_for_repo, repo):
+    try:
+        authed_get("verifying repository access", url_for_repo)
+    except NotFoundException:
+        # throwing user-friendly error message as it checks token access
+        message = "HTTP-error-code: 404, Error: Please check the repository name \'{}\' or you do not have sufficient permissions to access this repository.".format(repo)
+        raise NotFoundException(message) from None
+
+
+def verify_access_for_repo(config):
+
+    access_token = config['access_token']
+    session.headers.update({'authorization': 'token ' + access_token, 'per_page': '1', 'page': '1'})
+
+    repositories = list(filter(None, config['repository'].split(' ')))
+
+    for repo in repositories:
+        logger.info("Verifying access of repository: %s", repo)
+
+        url_for_repo = "https://api.github.com/repos/{}/commits".format(repo)
+
+        # Verifying for Repo access
+        verify_repo_access(url_for_repo, repo)
+
+
+# return the 'timeout'
+def get_request_timeout():
+    args = singer.utils.parse_args([])
+    # get the value of request timeout from config
+    config_request_timeout = args.config.get('request_timeout')
+
+    # only return the timeout value if it is passed in the config and the value is not 0, "0" or ""
+    if config_request_timeout and float(config_request_timeout):
+        # return the timeout from config
+        return float(config_request_timeout)
+
+    # return default timeout
+    return REQUEST_TIMEOUT
