+++ tap_github/issues/comments.py
@@ -1,0 +1,30 @@
+# Python imports
+# Third-Party imports
+import singer
+# Project imports
+
+
+def get_all_comments(schema, repo_path, state, mdata, start_date):
+    '''
+    https://developer.github.com/v3/issues/comments/#list-comments-in-a-repository
+    '''
+    bookmark = get_bookmark(state, repo_path, "comments", "since", start_date)
+    if bookmark:
+        query_string = '&since={}'.format(bookmark)
+    else:
+        query_string = ''
+    with metrics.record_counter('comments') as counter:
+        for response in authed_get_all_pages(
+                'comments',
+                'https://api.github.com/repos/{}/issues/comments?sort=updated&direction=asc{}'.format(repo_path, query_string)
+        ):
+            comments = response.json()
+            extraction_time = singer.utils.now()
+            for comment in comments:
+                comment['_sdc_repository'] = repo_path
+                with singer.Transformer() as transformer:
+                    rec = transformer.transform(comment, schema, metadata=metadata.to_map(mdata))
+                singer.write_record('comments', rec, time_extracted=extraction_time)
+                singer.write_bookmark(state, repo_path, 'comments', {'since': singer.utils.strftime(extraction_time)})
+                counter.increment()
+    return state
