+++ tap_github/issues/issues.py
@@ -1,0 +1,64 @@
+# Python imports
+# Third-Party imports
+import singer
+# Project imports
+
+
+def get_all_issues(schema, repo_path,  state, mdata, start_date):
+    '''
+    https://developer.github.com/v3/issues/#list-issues-for-a-repository
+    '''
+    bookmark = get_bookmark(state, repo_path, "issues", "since", start_date)
+    if bookmark:
+        query_string = '&since={}'.format(bookmark)
+    else:
+        query_string = ''
+
+    with metrics.record_counter('issues') as counter:
+        for response in authed_get_all_pages(
+                'issues',
+                'https://api.github.com/repos/{}/issues?state=all&sort=updated&direction=asc{}'.format(repo_path, query_string)
+        ):
+            issues = response.json()
+            extraction_time = singer.utils.now()
+            for issue in issues:
+                issue['_sdc_repository'] = repo_path
+                with singer.Transformer() as transformer:
+                    rec = transformer.transform(issue, schema, metadata=metadata.to_map(mdata))
+                singer.write_record('issues', rec, time_extracted=extraction_time)
+                singer.write_bookmark(state, repo_path, 'issues', {'since': singer.utils.strftime(extraction_time)})
+                counter.increment()
+    return state
+
+
+def get_all_issue_events(schemas, repo_path, state, mdata, start_date):
+    bookmark_value = get_bookmark(state, repo_path, "issue_events", "since", start_date)
+    if bookmark_value:
+        bookmark_time = singer.utils.strptime_to_utc(bookmark_value)
+    else:
+        bookmark_time = 0
+
+    with metrics.record_counter('issue_events') as counter:
+        for response in authed_get_all_pages(
+                'issue_events',
+                'https://api.github.com/repos/{}/issues/events?sort=created_at&direction=desc'.format(repo_path)
+        ):
+            events = response.json()
+            extraction_time = singer.utils.now()
+            for event in events:
+                event['_sdc_repository'] = repo_path
+                # skip records that haven't been updated since the last run
+                # the GitHub API doesn't currently allow a ?since param for pulls
+                # once we find the first piece of old data we can return, thanks to
+                # the sorting
+                updated_at = event.get('created_at') if event.get('updated_at') is None else event.get('updated_at')
+                if bookmark_time and singer.utils.strptime_to_utc(updated_at) < bookmark_time:
+                    return state
+
+                # transform and write release record
+                with singer.Transformer() as transformer:
+                    rec = transformer.transform(event, schemas, metadata=metadata.to_map(mdata))
+                singer.write_record('issue_events', rec, time_extracted=extraction_time)
+                singer.write_bookmark(state, repo_path, 'issue_events', {'since': singer.utils.strftime(extraction_time)})
+                counter.increment()
+    return state
