+++ tap_github/issues/labels.py
@@ -1,0 +1,28 @@
+# Python imports
+# Third-Party imports
+import singer
+# Project imports
+
+
+def get_all_issue_labels(schemas, repo_path, state, mdata, _start_date):
+    # https://developer.github.com/v3/issues/labels/
+    # not sure if incremental key
+    # 'https://api.github.com/repos/{}/labels?sort=created_at&direction=desc'.format(repo_path)
+    with metrics.record_counter('issue_labels') as counter:
+        for response in authed_get_all_pages(
+                'issue_labels',
+                'https://api.github.com/repos/{}/labels'.format(repo_path)
+        ):
+            issue_labels = response.json()
+            extraction_time = singer.utils.now()
+            for r in issue_labels:
+                r['_sdc_repository'] = repo_path
+
+                # transform and write release record
+                with singer.Transformer() as transformer:
+                    rec = transformer.transform(r, schemas, metadata=metadata.to_map(mdata))
+                singer.write_record('issue_labels', rec, time_extracted=extraction_time)
+                singer.write_bookmark(state, repo_path, 'issue_labels', {'since': singer.utils.strftime(extraction_time)})
+                counter.increment()
+    return state
+
