+++ tap_github/issues/milestones.py
@@ -1,0 +1,38 @@
+# Python imports
+# Third-Party imports
+import singer
+# Project imports
+
+
+def get_all_issue_milestones(schemas, repo_path, state, mdata, start_date):
+    # Incremental sync off `due on` ??? confirm.
+    # https://developer.github.com/v3/issues/milestones/#list-milestones-for-a-repository
+    # 'https://api.github.com/repos/{}/milestones?sort=created_at&direction=desc'.format(repo_path)
+    bookmark_value = get_bookmark(state, repo_path, "issue_milestones", "since", start_date)
+    if bookmark_value:
+        bookmark_time = singer.utils.strptime_to_utc(bookmark_value)
+    else:
+        bookmark_time = 0
+    with metrics.record_counter('issue_milestones') as counter:
+        for response in authed_get_all_pages(
+                'milestones',
+                'https://api.github.com/repos/{}/milestones?direction=desc'.format(repo_path)
+        ):
+            milestones = response.json()
+            extraction_time = singer.utils.now()
+            for r in milestones:
+                r['_sdc_repository'] = repo_path
+                # skip records that haven't been updated since the last run
+                # the GitHub API doesn't currently allow a ?since param for pulls
+                # once we find the first piece of old data we can return, thanks to
+                # the sorting
+                if bookmark_time and r.get("due_on") and singer.utils.strptime_to_utc(r.get("due_on")) < bookmark_time:
+                    continue
+
+                # transform and write release record
+                with singer.Transformer() as transformer:
+                    rec = transformer.transform(r, schemas, metadata=metadata.to_map(mdata))
+                singer.write_record('issue_milestones', rec, time_extracted=extraction_time)
+                singer.write_bookmark(state, repo_path, 'issue_milestones', {'since': singer.utils.strftime(extraction_time)})
+                counter.increment()
+    return state
