+++ tap_github/pull_requests/assignees.py
@@ -1,0 +1,25 @@
+# Python imports
+# Third-Party imports
+import singer
+# Project imports
+
+
+def get_all_assignees(schema, repo_path, state, mdata, _start_date):
+    '''
+    https://developer.github.com/v3/issues/assignees/#list-assignees
+    '''
+    with metrics.record_counter('assignees') as counter:
+        for response in authed_get_all_pages(
+                'assignees',
+                'https://api.github.com/repos/{}/assignees'.format(repo_path)
+        ):
+            assignees = response.json()
+            extraction_time = singer.utils.now()
+            for assignee in assignees:
+                assignee['_sdc_repository'] = repo_path
+                with singer.Transformer() as transformer:
+                    rec = transformer.transform(assignee, schema, metadata=metadata.to_map(mdata))
+                singer.write_record('assignees', rec, time_extracted=extraction_time)
+                singer.write_bookmark(state, repo_path, 'assignees', {'since': singer.utils.strftime(extraction_time)})
+                counter.increment()
+    return state
