+++ tap_github/pull_requests/comments.py
@@ -1,0 +1,18 @@
+# Python imports
+# Third-Party imports
+import singer
+# Project imports
+
+
+def get_review_comments_for_pr(pr_number, schema, repo_path, state, mdata):
+    for response in authed_get_all_pages(
+            'comments',
+            'https://api.github.com/repos/{}/pulls/{}/comments'.format(repo_path,pr_number)
+    ):
+        review_comments = response.json()
+        for comment in review_comments:
+            comment['_sdc_repository'] = repo_path
+            with singer.Transformer() as transformer:
+                rec = transformer.transform(comment, schema, metadata=metadata.to_map(mdata))
+            yield rec
+        return state
