+++ tap_github/pull_requests/pulls.py
@@ -1,0 +1,79 @@
+# Python imports
+# Third-Party imports
+import singer
+import singer.metrics as metrics
+from singer import metadata
+# Project imports
+from gh_client import get_bookmark, authed_get_all_pages, authed_get
+from pull_requests.pr_detail import get_pr_detail, enhance_pull
+from pull_requests.reviews import get_reviews_for_pr
+from pull_requests.comments import get_review_comments_for_pr
+from pull_requests.commits import get_commits_for_pr
+
+
+def get_all_pull_requests(schemas, repo_path, state, mdata, start_date):
+    '''
+    https://developer.github.com/v3/pulls/#list-pull-requests
+    '''
+
+    bookmark_value = get_bookmark(state, repo_path, "pull_requests", "since", start_date)
+    if bookmark_value:
+        bookmark_time = singer.utils.strptime_to_utc(bookmark_value)
+    else:
+        bookmark_time = 0
+    with metrics.record_counter('pull_requests') as counter:
+        with metrics.record_counter('reviews') as reviews_counter:
+            for response in authed_get_all_pages(
+                    'pull_requests',
+                    'https://api.github.com/repos/{}/pulls?state=all&sort=updated&direction=desc'.format(repo_path)
+            ):
+                pull_requests = response.json()
+                extraction_time = singer.utils.now()
+                for pr in pull_requests:
+                    # skip records that haven't been updated since the last run
+                    # the GitHub API doesn't currently allow a ?since param for pulls
+                    # once we find the first piece of old data we can return, thanks to
+                    # the sorting
+                    if bookmark_time and singer.utils.strptime_to_utc(pr.get('updated_at')) < bookmark_time:
+                        return state
+
+                    pr_num = pr.get('number')
+                    pr_id = pr.get('id')
+                    pr['_sdc_repository'] = repo_path
+                    pr_detail = get_pr_detail(pr_num, repo_path)
+                    pr = enhance_pull(pr, pr_detail)
+
+                    # transform and write pull_request record
+                    with singer.Transformer() as transformer:
+                        rec = transformer.transform(pr, schemas['pull_requests'], metadata=metadata.to_map(mdata))
+                    singer.write_record('pull_requests', rec, time_extracted=extraction_time)
+                    singer.write_bookmark(state, repo_path, 'pull_requests', {'since': singer.utils.strftime(extraction_time)})
+                    counter.increment()
+
+                    # sync reviews if that schema is present (only there if selected)
+                    if schemas.get('reviews'):
+                        for review_rec in get_reviews_for_pr(pr_num, schemas['reviews'], repo_path, state, mdata, pr_id):
+                            singer.write_record('reviews', review_rec, time_extracted=extraction_time)
+                            singer.write_bookmark(state, repo_path, 'reviews', {'since': singer.utils.strftime(extraction_time)})
+
+                            reviews_counter.increment()
+
+                    # sync review comments if that schema is present (only there if selected)
+                    if schemas.get('review_comments'):
+                        for review_comment_rec in get_review_comments_for_pr(pr_num, schemas['review_comments'], repo_path, state, mdata):
+                            singer.write_record('review_comments', review_comment_rec, time_extracted=extraction_time)
+                            singer.write_bookmark(state, repo_path, 'review_comments', {'since': singer.utils.strftime(extraction_time)})
+
+                    if schemas.get('pr_commits'):
+                        for pr_commit in get_commits_for_pr(
+                                pr_num,
+                                pr_id,
+                                schemas['pr_commits'],
+                                repo_path,
+                                state,
+                                mdata
+                        ):
+                            singer.write_record('pr_commits', pr_commit, time_extracted=extraction_time)
+                            singer.write_bookmark(state, repo_path, 'pr_commits', {'since': singer.utils.strftime(extraction_time)})
+
+    return state
