+++ tap_github/starrings.py
@@ -1,0 +1,27 @@
+# Python imports
+# Third-Party imports
+import singer
+# Project imports
+
+def get_all_stargazers(schema, repo_path, state, mdata, _start_date):
+    '''
+    https://developer.github.com/v3/activity/starring/#list-stargazers
+    '''
+    stargazers_headers = {'Accept': 'application/vnd.github.v3.star+json'}
+    with metrics.record_counter('stargazers') as counter:
+        for response in authed_get_all_pages(
+                'stargazers',
+                'https://api.github.com/repos/{}/stargazers'.format(repo_path), stargazers_headers
+        ):
+            stargazers = response.json()
+            extraction_time = singer.utils.now()
+            for stargazer in stargazers:
+                user_id = stargazer['user']['id']
+                stargazer['_sdc_repository'] = repo_path
+                with singer.Transformer() as transformer:
+                    rec = transformer.transform(stargazer, schema, metadata=metadata.to_map(mdata))
+                rec['user_id'] = user_id
+                singer.write_record('stargazers', rec, time_extracted=extraction_time)
+                singer.write_bookmark(state, repo_path, 'stargazers', {'since': singer.utils.strftime(extraction_time)})
+                counter.increment()
+    return state
