+++ tap_github/streams.py
@@ -1,0 +1,48 @@
+# Python imports
+# Third-Party imports
+import singer
+# Project imports
+
+
+def get_catalog():
+    raw_schemas = load_schemas()
+    streams = []
+    for schema_name, schema in raw_schemas.items():
+        # get metadata for each field
+        mdata = populate_metadata(schema_name, schema)
+        # create and add catalog entry
+        catalog_entry = {
+            'stream': schema_name,
+            'tap_stream_id': schema_name,
+            'schema': schema,
+            'metadata' : metadata.to_list(mdata),
+            'key_properties': KEY_PROPERTIES[schema_name],
+        }
+        streams.append(catalog_entry)
+    return {'streams': streams}
+
+
+def get_selected_streams(catalog):
+    '''
+    Gets selected streams.  Checks schema's 'selected'
+    first -- and then checks metadata, looking for an empty
+    breadcrumb and mdata with a 'selected' entry
+    '''
+    selected_streams = []
+    for stream in catalog['streams']:
+        stream_metadata = stream['metadata']
+        if stream['schema'].get('selected', False):
+            selected_streams.append(stream['tap_stream_id'])
+        else:
+            for entry in stream_metadata:
+                # stream metadata will have empty breadcrumb
+                if not entry['breadcrumb'] and entry['metadata'].get('selected', None):
+                    selected_streams.append(stream['tap_stream_id'])
+    return selected_streams
+
+
+def get_stream_from_catalog(stream_id, catalog):
+    for stream in catalog['streams']:
+        if stream['tap_stream_id'] == stream_id:
+            return stream
+    return None
